work in progress: added EN/NL with translation strings | cleaned index.html of hardcoded modal window | seperate file for modal screen rendering
