+++ data/core.yaml
@@ -1,4 +1,10 @@
 en:
+  add-image-modal:
+    addbutton: Add picture
+    title: Add a picture
+    submit: Upload picture
+    reminder: Reminder
+    license: Your picture will be uploaded online on Wikimedia Commons with the CC 1.0 license. This means your photo will be dedicated to the public domain.
   icons:
     download: download
     information: info
